requirements.txt
Flask==3.0.0
PyMuPDF==1.23.8
sentence-transformers==3.0.1
faiss-cpu==1.7.4
transformers==4.40.0
torch==2.0.1
torchvision==0.15.2
accelerate==0.25.0
Werkzeug==3.0.1
numpy==1.26.4
huggingface-hub==0.25.2
tokenizers==0.19
protobuf==4.24.0
sentencepiece==0.1.99